// gen/models/group.ts
// tslint:disable
/**
 * 实训打卡
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface Group
 */
export interface Group {
    /**
     * 用户组ID
     * @type {number}
     * @memberof Group
     */
    groupId?: number;
    /**
     * 用户组名称
     * @type {string}
     * @memberof Group
     */
    groupName?: string;
    /**
     * 用户组描述
     * @type {string}
     * @memberof Group
     */
    description?: string;
    /**
     * 创建者用户ID
     * @type {number}
     * @memberof Group
     */
    creatorId?: number;
    /**
     * 创建者用户名
     * @type {string}
     * @memberof Group
     */
    creatorName?: string;
    /**
     * 成员数量
     * @type {number}
     * @memberof Group
     */
    memberCount?: number;
    /**
     * 创建时间（Unix时间戳，单位：秒）
     * @type {number}
     * @memberof Group
     */
    createdAt?: number;
}