// gen/models/group-member.ts
// tslint:disable
/**
 * 实训打卡
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface GroupMember
 */
export interface GroupMember {
    /**
     * 用户ID
     * @type {number}
     * @memberof GroupMember
     */
    userId?: number;
    /**
     * 用户名
     * @type {string}
     * @memberof GroupMember
     */
    username?: string;
    /**
     * 用户在组中的角色，如\'admin\'或\'member\'
     * @type {string}
     * @memberof GroupMember
     */
    role?: string;
    /**
     * 加入时间（Unix时间戳，单位：秒）
     * @type {number}
     * @memberof GroupMember
     */
    joinedAt?: number;
}